Drop various unused exported methods.
