+++ common/debug.go
@@ -21,7 +21,6 @@ import (
 	"os"
 	"runtime"
 	"runtime/debug"
-	"strings"
 )
 
 // Report gives off a warning requesting the user to submit an issue to the github tracker.
@@ -35,18 +34,4 @@ func Report(extra ...interface{}) {
 	debug.PrintStack()
 
 	fmt.Fprintln(os.Stderr, "#### BUG! PLEASE REPORT ####")
-}
-
-// PrintDepricationWarning prinst the given string in a box using fmt.Println.
-func PrintDepricationWarning(str string) {
-	line := strings.Repeat("#", len(str)+4)
-	emptyLine := strings.Repeat(" ", len(str))
-	fmt.Printf(`
-%s
-# %s #
-# %s #
-# %s #
-%s
-
-`, line, emptyLine, str, emptyLine, line)
 }
