+++ p2p/server.go
@@ -32,7 +32,7 @@ import (
 )
 
 const (
-	defaultDialTimeout      = 15 * time.Second
+	defaultDialTimeout = 15 * time.Second
 
 	// Maximum number of concurrently handshaking inbound connections.
 	maxAcceptConns = 50
